# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RegisAI Compliance Platform API
  version: 0.1.0
  description: |
    AI-powered compliance analysis platform. Upload compliance documentation,
    run gap analysis against regulatory frameworks, draft remediation policy
    language, and monitor regulatory changes.

    **Authentication:** All endpoints require a valid Supabase session cookie
    (set via /login or /auth/callback). Unauthenticated requests return 401.

    **Data isolation:** Row-Level Security enforces per-user data isolation.
    Users can only access their own documents, audits, and findings.
  contact:
    name: RegisAI Team
    url: https://github.com/bitcoinow/RegisAI
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://regis-ai-nine.vercel.app
    description: Production (Vercel)
...
paths:
  /api/documents:
    get:
      tags: [Documents]
      summary: List documents
      description: Returns all documents for the authenticated user, ordered by creation date descending.
      operationId: listDocuments
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [Documents]
      summary: Upload PDF document
      description: |
        Accepts a PDF file via multipart form data. Extracts text using pdf-parse,
        uploads the original PDF to Supabase Storage, and creates a document record.
        Only PDF files are accepted (415 for other types).
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file (max 10MB)
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUploadResponse"
        "400":
          description: Missing file
        "401":
          $ref: "#/components/responses/Unauthorized"
        "415":
          description: Not a PDF file
        "422":
          description: PDF text extraction failed
...
components:
...
  responses:
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
    DocumentSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        file_name:
          type: string
        page_count:
          type: integer
          nullable: true
        status:
          type: string
          enum: [uploaded, analysing, complete, error]
        created_at:
          type: string
          format: date-time
    DocumentUploadResponse:
      type: object
      properties:
        document_id:
          type: string
          format: uuid
        page_count:
          type: integer
...